Rename classes in NCalcPlugins
